Infrastructure (grunt tasks) for es6 module transpilation via esperanto.
This commit sets the groundwork for modularizing the code but does not actually add dependencies to individual files,
those will be merged into this branch in subsequent commits. Thus the new build artifacts in /dist don't actually work
at the moment as they will refer to a number of things that are not out of scope until they are explicitly imported.

This commit also includes some changes to SimpleGraphics to be strict mode compliant so that it can be transpiled.