Changes required to get the List module to be loadable.
Note that just because it loads doesn't mean that every function in it would work.

A few changes to the build process are also included, we are using esperanto directly (rather than via the grunt
plugin), and allowing it to all the building and concatenating itself in order to properly handle circular dependencies.
In this mode the tool itself traces dependencies starting from index.js.