Adds @todo JSDoc tag to functions with no docs.
This commit adds todo tags for the namespace
objects like operators and converters.

This is the first step towards understanding the
scope left in terms of documenting the MO code.

The process adds the function name to the
generated JSDocs - but without additional comments.

We can search for this @todo tag easily to
determine areas of under-documentation.